==> notebooks/08_test_pipeline.ipynb <==
import os
import sys
import pandas as pd
from imblearn.over_sampling import RandomOverSampler
sys.path.append(os.path.abspath(os.path.join('..', 'src',)))

from models.classifiers.ridge import CustomRidge
from models.embeddings.openai_embedding import OpenAIEmbedding



embedding = OpenAIEmbedding(model_name='text-embedding-3-small')

outputs_df = pd.read_parquet("../data/processed/outputs_openai_embeddings_v1.parquet")
ppas = outputs_df.PPAs_list.explode().drop_duplicates().sort_values().to_list()
# train_df = outputs_df[~outputs_df.test_set]
# test_df = outputs_df[outputs_df.test_set]
X_labels = outputs_df.columns[outputs_df.columns.str.contains("openai_embedding_small")]
# y_label = "PPAs_list"
oversampling = True

X_train = outputs_df[X_labels].to_numpy()
# y_train = train_df[y_label].to_numpy()
y_train = outputs_df[ppas].astype(int).to_numpy()

# X_test = test_df[X_labels].to_numpy()
# # y_test = test_df[y_label].to_numpy()
# y_test = test_df[ppas].astype(int).to_numpy()


ros = RandomOverSampler(random_state=42)
train_oversamples_df, _ = ros.fit_resample(outputs_df, y=outputs_df["primary_ppa"])
X_train_oversampled = train_oversamples_df[X_labels].to_numpy()
# y_train = train_df[y_label].to_numpy()
y_train_oversampled = train_oversamples_df[ppas].astype(int).to_numpy() 


outputs_df[ppas].sum().sort_values(ascending=False)


ridge_clf = CustomRidge()
ridge_clf.fit(X_train_oversampled, y_train_oversampled)


import pickle

# Pickle the ridge_clf model
with open('../data/classifiers/ridge_clf.pkl', 'wb') as f:
    pickle.dump(ridge_clf, f)



# Unpickle the ridge clf model
with open('../data/classifiers/ridge_clf.pkl', 'rb') as f:
    ridge_clf_1 = pickle.load(f)

ridge_clf_1.classes_


ppas

ppasappas

e1 = embedding.generate("Under this Programme Priority Area (PPA), FAO will support national and local decision-makers to initiate, coordinate and scale up actions and investments towards urban and peri-urban food systems transformation to (i) reduce the level of urban poverty, food insecurity and malnutrition; (ii) address gender and social inequalities in cities and towns through improved urban and territorial food-policy, planning and integrated actions and inclusive governance; (iii) support an inclusive and sustainable rural transformation by leveraging urban demand for healthy diets and improved rural–urban and territorial linkages; (iv) promote sustainable management of natural resources (including natural ecosystems, green spaces, water, land, energy, and wild foods); (v) improve urban and territorial resilience of food systems to climatic and other shocks such as chronic stresses and extreme events; and (vi) strengthen partnerships at all levels and promote urban food systems alliances and coalitions with relevant actors.")

pd.DataFrame(ridge_clf.predict_proba(e1.reshape(1, -1)), columns=ppas).iloc[0].idxmax()
