==> notebooks/10_predict_backup.ipynb <==
import os
import sys
import pandas as pd
import pickle
sys.path.append(os.path.abspath(os.path.join('..', 'src',)))
from models.embeddings.openai_embedding import OpenAIEmbedding

embedding = OpenAIEmbedding()

ppas = ['BE.1',
        'BE.2',
        'BE.3',
        'BE.4',
        'BL.1',
        'BL.2',
        'BL.3',
        'BL.4',
        'BL.5',
        'BL.6',
        'BN.1',
        'BN.2',
        'BN.3',
        'BN.4',
        'BN.5',
        'BP.1',
        'BP.2',
        'BP.3',
        'BP.4',
        'BP.5']


with open('../data/classifiers/ridge_clf.pkl', 'rb') as f:
    clf = pickle.load(f)

df1 = pd.read_excel("../data/raw/WP sugerencia de outputs.xlsx", sheet_name="BACKUP")

df2 = pd.read_excel("../data/raw/WP sugerencia de outputs.xlsx", sheet_name="SELECCION")

df_to_predict = df1[~df1["Output Statement"].isin(df2["Output Statement"])]

df_to_predict["embedding_small"] = df_to_predict["Output Statement"].apply(lambda x: embedding.generate(x))

df_to_predict.to_parquet("../data/processed/backup_embeddings.parquet")

df_to_predict

import numpy as np


X_backup = np.array(df_to_predict["embedding_small"].tolist())
X_backup.shape

y_pred_backup = clf.predict_proba(X_backup)
y_pred_backup


y_pred_backup_df = pd.DataFrame(y_pred_backup, columns=ppas)
y_pred_backup_df["predicted_ppas"] = y_pred_backup_df.apply(lambda x: x.sort_values(ascending=False).head(3).index.to_list(), axis=1)

df_to_predict.reset_index(drop=True).join(y_pred_backup_df[["predicted_ppas"]]).to_parquet("../data/processed/backup_embeddings_predictions_v1.parquet")

df_to_predict_predicted = (df_to_predict
 .reset_index(drop=True)
 .drop(columns=["embedding_small"])
 .join(y_pred_backup_df[["predicted_ppas"]]))

df_to_predict_predicted

df_to_predict_predicted["predicted_intersection"] = df_to_predict_predicted.apply(lambda row: len(set(row["PPAs"].split(", ")).intersection(set(row["predicted_ppas"]))), axis=1)

df_to_predict_predicted.to_excel("../data/processed/backup_predictions_v2.xlsx", index=False)

training_set = pd.read_parquet("../data/processed/outputs_openai_embeddings_v1.parquet")

v1 = training_set[~training_set["test_set"]].PPAs.str.split(", ").explode().value_counts()
v2 = df_to_predict_predicted[df_to_predict_predicted["predicted_intersection"].isin([3, 2])][["PPAs", "predicted_ppas"]].PPAs.str.split(", ").explode().value_counts()

(v1 + v2).sort_values(ascending=False)
